Address linting errors for interpreter files
