+++ lib/interpreter/interpreter.ts
@@ -5,55 +5,29 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { AstNode, BinaryOperatorNode, UnaryOperatorNode, ValueOperandNode } from '../parser/ast';
 import { TokenType } from "../lexer/token";
-import { Logger } from '../util/logger';
-
-export class NodeVisitor {
-
-  visit(node:AstNode):boolean {
-    if (!node) {
-      return null;
-    }
-
-    let visitorName:string = this.getMethodName(node);
-
-    Logger.log(`Visiting: ${node} with method: ${visitorName}`);
-
-    if (this[visitorName]) {
-      return this[visitorName](node);
-    } else {
-      return this.genericVisit(node);
-    }
-  }
-
-  private genericVisit(node:AstNode):boolean {
-    throw `No visitor method: ${this.getMethodName(node)}`;
-  }
-
-  private getMethodName(node:AstNode):string {
-    return `visit${node.name}`;
-  }
-}
+import { AstNode, BinaryOperatorNode, UnaryOperatorNode, ValueOperandNode } from "../parser/ast";
+import { Logger } from "../util/logger";
+import { NodeVisitor } from "./node_visitor";
 
 export class Interpreter extends NodeVisitor {
-  readonly resolver:Function;
+  private readonly resolver: () => void;
 
-  constructor(resolver:Function) {
+  constructor(resolver: () => void) {
     super();
 
     if (!resolver) {
-      throw 'Resolver is required'
+      throw new Error("Resolver is required");
     }
 
     this.resolver = resolver;
   }
 
-  private error(node:AstNode):boolean {
-    throw `Visitor cant process node token type: ${node.token.type}`;
+  private error(node: AstNode): boolean {
+    throw new Error(`Visitor cant process node token type: ${node.token.type}`);
   }
 
-  private visitBinaryOperatorNode(node:BinaryOperatorNode):boolean {
+  private visitBinaryOperatorNode(node: BinaryOperatorNode): boolean {
     if (node.token.type === TokenType.AndOp) {
       return this.visit(node.left) && this.visit(node.right);
     } else if (node.token.type === TokenType.OrOp) {
@@ -63,7 +37,7 @@ export class Interpreter extends NodeVisitor {
     }
   }
 
-  private visitUnaryOperatorNode(node:UnaryOperatorNode):boolean {
+  private visitUnaryOperatorNode(node: UnaryOperatorNode): boolean {
     if (node.token.type === TokenType.NotOp) {
       return !this.visit(node.child);
     } else {
@@ -71,18 +45,18 @@ export class Interpreter extends NodeVisitor {
     }
   }
 
-  private visitValueOperandNode(node:ValueOperandNode):boolean {
-    if (node.value === 'true') {
+  private visitValueOperandNode(node: ValueOperandNode): boolean {
+    if (node.value === "true") {
       return true;
-    } else if (node.value === 'false' || node.value === 'null') {
+    } else if (node.value === "false" || node.value === "null") {
       return false;
     } else {
       return this.resolveValue(node.value);
     }
   }
 
-  private resolveValue(value:string) {
-    let resolvedValue = !!this.resolver(value);
+  private resolveValue(value: string) {
+    const resolvedValue = !!this.resolver(value);
 
     Logger.log(`Resolved: ${value} to: ${resolvedValue}`);
